Move the password generation helper to the services class
