+++ src/EventManagement.Services/TalentLms/ITalentLmsService.cs
@@ -1,5 +1,4 @@
 using System.Threading.Tasks;
-using EventManagement.Services.Extensions;
 using losol.EventManagement.Services.TalentLms.Models;
 using Microsoft.Extensions.Options;
 using RestSharp;
